requirements.txt
pyautogui>=0.9.54
Pillow>=12.0.0
numpy>=2.0.0
opencv-python>=4.11.0
pywin32>=311
